requirements_portfolio.txt
yfinance==0.2.18
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0